# Create prompt for AI coding 🚀 Complete Migration & Enhancement: Crypto Analysis Suite v2.0 assistant