--- a/src/vite-env.d.ts
+++ b/src/vite-env.d.ts
@@ -1,0 +1,10 @@
+/// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  readonly VITE_GEMINI_API_KEY: string
+  // más variables de entorno...
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}