… CHECKLIST_VERIFICACION.md
… CREATE_PULL_REQUEST_WITH_CLAUDE_CLI.md
… Claude.Desktop.Instructions/0. ai_coding_assistant_prompt.md
… Claude.Desktop.Instructions/0. background_agent_checklist.md
… Claude.Desktop.Instructions/0. background_agent_instructions.md
… Claude.Desktop.Instructions/0. commit_strategy.md
… Claude.Desktop.Instructions/0. environment_setup_guide.md
… Claude.Desktop.Instructions/0. package_json_template.json
… Claude.Desktop.Instructions/0. pull_request_template.md
… Claude.Desktop.Instructions/1. migration_guide.md
… Claude.Desktop.Instructions/13. heatmap_config_fix.ts
… Claude.Desktop.Instructions/3. instrucciones_csv_integration.md
… Claude.Desktop.Instructions/4. csv_file_handler.ts
… Claude.Desktop.Instructions/5. sector_heatmap_csv_integration.ts
… Claude.Desktop.Instructions/6. csv_integration_service.ts
… Claude.Desktop.Instructions/7. plan_implementacion.md
… Claude.Desktop.Instructions/8. estadisticas_dinamicas.ts
… Claude.Desktop.Instructions/9. paginas_dinamicas.ts
… CodingAgentPreviousChat.md
… CryptoAnalysisExports/README.md
… CryptoAnalysisExports/Template_DDMMYYYY_Thh:mm.md.md
… CryptoAnalysisExports/Template_example_DDMMYYYY_Thh:mm.md
… CryptoAnalysisExports/crypto_analysis_template.json
… CryptoAnalysisExports/crypto_coins_category.csv
… CryptoAnalysisExports/crypto_market_analysis_2025-07-05T01-39-13.md
… CryptoAnalysisExports/crypto_market_analysis_2025-07-05T02-03-32.md
… CryptoAnalysisExports/crypto_market_analysis_2025-07-05T02-13-09.md
… CryptoAnalysisExports/crypto_market_analysis_2025-07-05T03-07-50.md
… GeminiInstruccions.md
… IMPLEMENTACION_COMPLETADA.md