@Nightmurder97 2025-07-05
## 🔄 CI/CD Workflow Update Required

The new GitHub Actions workflow () needs to be added to this PR for automated testing.

**Missing CI/CD Coverage:**
- Build validation for Node.js 18.x & 20.x
- ESLint configuration (currently missing)
- TypeScript compilation checks
- Security audit (npm audit)

**Recommendation:** Add the CI/CD workflow to ensure quality checks run on every PR.